requirements.txt
Django==4.2.1
Pillow==9.5.0
Pillow==9.5.0
pypdf==3.9.1
python-dotenv==1.0.0
telepot==12.7
